requirements.txt
absl-py==0.9.0
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
gast==0.3.3
gin-config==0.3.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.1
Pillow==7.2.0
protobuf==3.12.2
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
matplotlib==3.3.1
